Add files via upload
This allows the use of the flet scenes to set up the db, connect the db to flet scnes to allow sign up, and login. It also includes some code to open the database and check to see if the chagnes went through as well. TLDR: Encrypted database setup works, Sign up works, Login works, program to access and inspect db included (file named: tamper_with_db.py) .

All code but vendor2.py & transaction.py were tested. The code was tested on windows 10. None of this code has been tested on Mac or Linux. NONE OF THIS HAS BEEN TESTED ON MAC OR LINUX. It should work on mac os as long as the libraries are properly installed.

 (Must have the proper libraries installed)

This has also been uploaded to my branch as a backup. We should build off this code going forward.